requirements.txt
requests~=2.31.0
beautifulsoup4~=4.12.2
aiohttp~=3.8.5
tqdm~=4.65.0
pyyaml~=6.0